Chase changes in builtin return types.

+++ sys/cheri/cheric.h
@@ -247,8 +247,8 @@ cheri_bytes_remaining(const void * __capability cap)
 
 #define _CHERI_PRINTF_CAP_FMT  "v:%lu s:%lu p:%08lx b:%016jx l:%016zx o:%jx t:%ld"
 #define _CHERI_PRINTF_CAP_ARG(ptr)					\
-	    cheri_gettag((const void * __capability)(ptr)),		\
-	    cheri_getsealed((const void * __capability)(ptr)),		\
+	    (unsigned long)cheri_gettag((const void * __capability)(ptr)),		\
+	    (unsigned long)cheri_getsealed((const void * __capability)(ptr)),		\
 	    cheri_getperm((const void * __capability)(ptr)),		\
 	    cheri_getbase((const void * __capability)(ptr)),		\
 	    cheri_getlen((const void * __capability)(ptr)),		\
